refactor: models, views, serializer file sistem
