--- metal_ops/serializers.py
@@ -1,36 +1,0 @@
-from rest_framework import serializers
-from .models import Sede, Cliente, OrdenTrabajo, Maquina, Servicio, Tarea
-
-class SedeSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Sede
-        fields = '__all__'
-
-
-class ClienteSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Cliente
-        fields = '__all__'
-
-
-class OrdenTrabajoSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = OrdenTrabajo
-        fields = '__all__'
-
-
-class MaquinaSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Maquina
-        fields = '__all__'
-
-
-class ServicioSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Servicio
-        fields = '__all__'
-
-class TareaSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Tarea
-        fields = '__all__'
