+++ metal_ops/serializers/__init__.py
@@ -1,0 +1,6 @@
+from .sede import SedeSerializer
+from .cliente import ClienteSerializer
+from .orden_trabajo import OrdenTrabajoSerializer
+from .maquina import MaquinaSerializer
+from .servicio import ServicioSerializer
+from .tarea import TareaSerializer
