+++ metal_ops/serializers/cliente.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+from metal_ops.models import Cliente
+
+
+class ClienteSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Cliente
+        fields = '__all__'
