+++ metal_ops/serializers/orden_trabajo.py
@@ -1,0 +1,7 @@
+from rest_framework import serializers
+from metal_ops.models import OrdenTrabajo
+
+class OrdenTrabajoSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = OrdenTrabajo
+        fields = '__all__'
