+++ metal_ops/serializers/sede.py
@@ -1,0 +1,7 @@
+from rest_framework import serializers
+from metal_ops.models import Sede
+
+class SedeSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Sede
+        fields = '__all__'
