+++ metal_ops/serializers/servicio.py
@@ -1,0 +1,7 @@
+from rest_framework import serializers
+from metal_ops.models import Servicio
+
+class ServicioSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Servicio
+        fields = '__all__'
