+++ metal_ops/serializers/tarea.py
@@ -1,0 +1,7 @@
+from rest_framework import serializers
+from metal_ops.models import Tarea
+
+class TareaSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Tarea
+        fields = '__all__'
