+++ metal_ops/views/__init__.py
@@ -1,0 +1,14 @@
+from .sede import CrearSedeView, ListarSedeView
+from .cliente import (
+    CrearClienteView, ListarClienteView, EditarClienteView, EliminarClienteView
+)
+from .orden_trabajo import (
+    CrearOrdenTrabajoView, ListarOrdenTrabajoView,
+    EditarOrdenTrabajoView, EliminarOrdenTrabajoView
+)
+from .maquina import CrearMaquinaView, ListarMaquinaView
+from .servicio import CrearServicioView, ListarServicioView
+from .tarea import (
+    CrearTareaView, ListarTareaView,
+    EditarTareaView, EliminarTareaView
+)
